requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
kafka-python
playwright
python-dotenv
pydantic
aiogram
